Handle invalidated pollfd references
When holding pollfd's in a vector to pass to poll,
it's quick and easy to remove and add pollfds in connect/disconnect.

But holding pointers to these in Peer::Connection is UB.
When erasing from the vector, the references are invalidated.
  i.e. each peer will hold a pointer to the position of the pollfd
  this means after an erase, Peer_Map::Connections hold pointers to the
  wrong or non-existent pollfd's

So we hold the pollfd's in Peer::Connection to simplify the execution
pipeline (i.e send in a peer that holds everything, each stage works on
a peer). This allows us to decouple the socket from the rest of the
pipeline (each stage can still get at the socket if it needs to).

Copy the pollfd's out of the Peer::Connections into a vector to be
processed by poll();